port benchmark to geometry branch
